Add slot struct
